Fixed a real*4 vs real*8 bug that was causing rate coefficients to be nonsense.
git-svn-id: http://dminfo:8686/svn/repository/models/trunk@827 cdc4813b-4270-ec4f-936e-925f93a782c2

+++ dsm2/src/fixed/process_rate_coef.f
@@ -18,7 +18,7 @@
       integer :: groupno
       integer :: istat
       integer, external :: name_to_objno
-      real*4  :: coefficient_value
+      real*8  :: coefficient_value
       integer,external :: rate_variable_code
       integer,external :: ncc_code
